switch from WIND to CAS
per PMT #95714.

All the existing stand hostnames are registered. Unfortunately, they
still won't do wildcards, so anyone creating a new stand won't be able
to log into it until we send in the request. Forest doesn't seem to be
heavily used anymore though so I guess that's not a high priority.

After this goes through, I'll probably add a bit that just warns the
user creating a new stand that it needs to be registered with CAS, and
then sends an email to ccnmtl-sysadmin so we can put in the request. I
guess we could also get fancy and send it directly to Dan plexus-style,
but I'm not sure that's really worth it given how infrequently new
stands are now created.