:arrow_up: django 1.11
This app was still on django 1.10 because of some failing tests that
weren't straightforward to fix. Because Forest depends heavily on
hostnames, the tests were doing some particular things with that.

Mainly, I just needed to add all the test hostnames to ALLOWED_HOSTS
during the test runs.
https://docs.djangoproject.com/en/2.1/ref/settings/#allowed-hosts

I've also removed less because it's not being used.